`Iris`: Add HTTP and WebSocket clients for chat sessions and messages   (#474)
* Add APIClient & Common to dependencies

* Add IrisSessionCountDTO

* Add IrisChatHttpService

* Add IrisWebsocketServicce

* Move from Int64 to Int, IrisWebsocketService: Update docs, IrisWebsocketServiceImpl: Add Task.isCancelled check to avoid dangling streams

+++ ArtemisKit/Sources/Iris/Models/IrisCitationMetaDTO.swift
@@ -6,9 +6,9 @@
 //
 
 struct IrisCitationMetaDTO: Codable, Hashable {
-    let entityId: Int64
+    let entityId: Int
     let lectureTitle: String
     let lectureUnitTitle: String
-    let lectureId: Int64
-    let courseId: Int64
+    let lectureId: Int
+    let courseId: Int
 }
