+++ ArtemisKit/Sources/Iris/Models/IrisSessionDTO.swift
@@ -10,10 +10,10 @@ import Foundation
 /// Slim session shape returned by `/chat/{courseId}/sessions/overview`,
 /// used to render the chat-history list. Does **not** contain messages.
 struct IrisSessionDTO: Codable, Hashable, Identifiable {
-    let id: Int64
+    let id: Int
     let title: String?
     let creationDate: Date
     let mode: ChatServiceMode
-    let entityId: Int64
+    let entityId: Int
     let entityName: String?
 }
